# ai-assistant/bizaira-ai-assistant-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: generate
    description: AI content generation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /generate/generate-text:
    post:
      operationId: generateText
      tags: [generate]
      summary: Generate text content
      description: Generate AI-powered text content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  description: The text prompt for generation
                systemPrompt:
                  type: string
                  description: Optional system prompt to customize AI behavior
              required:
                - prompt
      responses:
        "200":
          description: Text generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
                required:
                  - text
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
  /generate/generate-image:
    post:
      operationId: generateImage
      tags: [generate]
      summary: Generate image content
      description: Generate AI-powered image content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
                  description: The image prompt for generation
              required:
                - prompt
      responses:
        "200":
          description: Image generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  imageUrl:
                    type: string
                required:
                  - imageUrl
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status